Upgrades to use Tomcat 8.0.53
Tomcat 8.0.53 is now used as it is fully compatible with Java 8.

In addition, application and database versions have been updated to
the latest version avaliable, and the Dockerfile has been reordered
to run VEP task at the beginning, as it takes long time to download.
This reduces the need to reexecute this task when modifying the
Dockerfile.