Fixes execution with existent database
The execution was failing when using an existing database due to two
reasons: i) missing /var/run/mysqld directory and ii) no permissions
for MySQL on the database directory.

This commit fixes this errors by modifying the run.sh script, which
now checks for this problems and solves them if required.